fixtures/python-app/requirements.txt
stripe==7.9.0
boto3==1.34.0
openai==1.12.0
requests==2.31.0
httpx==0.27.0
psycopg2-binary==2.9.9
redis==5.0.1
twilio==8.13.0
sendgrid==6.11.0
sentry-sdk==1.40.0
